requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
PyJWT
psycopg2-binary